+++ src/main/kotlin/com/github/hechtcarmel/jetbrainsindexmcpplugin/tools/navigation/FindUsagesTool.kt
@@ -66,7 +66,7 @@ class FindUsagesTool : AbstractMcpTool() {
         }
     }
 
-    override suspend fun execute(project: Project, arguments: JsonObject): ToolCallResult {
+    override suspend fun doExecute(project: Project, arguments: JsonObject): ToolCallResult {
         val file = arguments[ParamNames.FILE]?.jsonPrimitive?.content
             ?: return createErrorResult(ErrorMessages.missingRequiredParam(ParamNames.FILE))
         val line = arguments[ParamNames.LINE]?.jsonPrimitive?.int